repair _add_in_group(), _detach_from_group()

+++ gui/extensions.py
@@ -1450,39 +1450,43 @@ class GroupView(gtk.Alignment, EventSource):
         self.emit_event("delete-group", self.group)
 
     def _add_in_group(self):
-        for source in self.app.sources_repository.get_sources(
-                lambda s : s.picked):
-            self.group.add(source)
-            self.app.sources_repository.remove(source)
-        try:
-            self.view.add_source(self.group)
-        except:
-            print("Open tab view")
+        if self.tabview is not None:
+            for name, item in self.view.group_views:
+                self.view.remove(item)
+            for source in self.app.sources_repository.get_sources(
+                    lambda s : s.picked):
+                self.group.add(source)
+                self.app.sources_repository.remove(source)
+                self.view.add_source()
 
     def _detach_from_group(self):
-        window = gtk.Window(gtk.WINDOW_TOPLEVEL)
-        window.set_title("Group sources")
-        window.set_size_request(230, 30)
-        frame = gtk.Frame()
-        frame.show()
+        if self.tabview is not None:
+            window = gtk.Window(gtk.WINDOW_TOPLEVEL)
+            window.set_title("Group sources")
+            window.set_size_request(230, 30)
+            frame = gtk.Frame()
+            frame.show()
 
-        self.store = gtk.ListStore(str, object)
-        self.store.append(["Select source", None])
-        for source in self.group._sources:
-            self.store.append([os.path.basename(source.name), source])
-        self.combo1 = gtk.ComboBox(self.store)
-        cell = gtk.CellRendererText()
-        self.combo1.pack_start(cell, True)
-        self.combo1.add_attribute(cell, 'text', 0)
-        self.combo1.set_active(0)
-        self.combo1.connect("changed", self._detach_source)
-        self.combo1.show()
+            self.store = gtk.ListStore(str, object)
+            self.store.append(["Select source", None])
+            for source in self.group._sources:
+                self.store.append([os.path.basename(source.name), source])
+            self.combo1 = gtk.ComboBox(self.store)
+            cell = gtk.CellRendererText()
+            self.combo1.pack_start(cell, True)
+            self.combo1.add_attribute(cell, 'text', 0)
+            self.combo1.set_active(0)
+            self.combo1.connect("changed", self._detach_source)
+            self.combo1.show()
 
-        frame.add(self.combo1)
-        window.add(frame)
-        window.show()
+            frame.add(self.combo1)
+            window.add(frame)
+            window.show()
 
     def _detach_source(self, w):
+        for name, item in self.view.group_views:
+            if item in self.view.get_children():
+                self.view.remove(item)
         iter = w.get_active_iter()
         text = w.get_active_text()
         if iter != None:
@@ -1493,10 +1497,7 @@ class GroupView(gtk.Alignment, EventSource):
                 self.group.remove(source)
                 self.app.sources_repository.add(source)
                 self.combo1.set_active(0)
-                try:
-                    self.view.detach_source(source)
-                except:
-                    print("Open tabview")
+                self.view.detach_source()
 
     def _cb_unpack_group(self, group):
         for source in group:
